move hooks to react utils file
