+++ src/app/content/highlights/components/HighlightsPopUp.tsx
@@ -7,9 +7,9 @@ import * as authSelect from '../../../auth/selectors';
 import { User } from '../../../auth/types';
 import ScrollLock from '../../../components/ScrollLock';
 import { isHtmlElement } from '../../../guards';
+import { onEsc } from '../../../reactUtils';
 import theme from '../../../theme';
 import { AppState, Dispatch } from '../../../types';
-import { onEsc } from '../../../utils';
 import { closeMyHighlights } from '../actions';
 import * as selectors from '../selectors';
 import * as Styled from './HighlightStyles';
