+++ src/app/reactUtils.ts
@@ -1,4 +1,4 @@
-import { FocusEvent, HTMLElement } from '@openstax/types/lib.dom';
+import { FocusEvent, HTMLElement, KeyboardEvent } from '@openstax/types/lib.dom';
 import React from 'react';
 import { addSafeEventListener, elementDescendantOf } from './domUtils';
 import { isElement } from './guards';
@@ -34,4 +34,43 @@ export const onFocusLostHandler = (ref: React.RefObject<HTMLElement>, isEnabled:
 
 export const useFocusLost = (ref: React.RefObject<HTMLElement>, isEnabled: boolean, cb: () => void) => {
   React.useEffect(onFocusLostHandler(ref, isEnabled, cb), [ref, isEnabled]);
+};
+
+/**
+ * This function will return array where first item is a function which will set
+ * event listener for given element and second item is a function which will remove
+ * this listener.
+ *
+ * This function can be used in React class components.
+ */
+export const onEsc = (
+  element: HTMLElement, cb: () => void
+): [() => void, () => void] => {
+  const handler = (e: KeyboardEvent) => {
+    if (e.key === 'Escape') {
+      e.stopPropagation();
+      cb();
+    }
+  };
+
+  return [
+    () => element.addEventListener('keydown', handler),
+    () => element.removeEventListener('keydown', handler),
+  ];
+};
+
+export const onEscHandler = (element: React.RefObject<HTMLElement>, isEnabled: boolean, cb: () => void) => () => {
+  const el = element && element.current;
+  if (!el) { return; }
+
+  const [addEvListener, removeEvListener] = onEsc(el, cb);
+  if (isEnabled) {
+    addEvListener();
+  }
+
+  return removeEvListener;
+};
+
+export const useOnEsc = (element: React.RefObject<HTMLElement>, isEnabled: boolean, cb: () => void) => {
+  React.useEffect(onEscHandler(element, isEnabled, cb), [element, isEnabled]);
 };
