+++ src/app/utils.ts
@@ -1,9 +1,8 @@
+import { Document } from '@openstax/types/lib.dom';
 import React, { Ref } from 'react';
 import { getType } from 'typesafe-actions';
 import Sentry from '../helpers/Sentry';
 import { recordError } from './errors/actions';
-
-import { Document, HTMLElement, KeyboardEvent } from '@openstax/types/lib.dom';
 import { isPlainObject } from './guards';
 import {
   ActionHookBody,
@@ -173,42 +172,3 @@ export const merge = <T1 extends {}, T2 extends {}>(thing1: T1, thing2: T2): T1 
     ),
   }), {}),
 });
-
-/**
- * This function will return array where first item is a function which will set
- * event listener for given element and second item is a function which will remove
- * this listener.
- *
- * This function can be used in React class components.
- */
-export const onEsc = (
-  element: HTMLElement, cb: () => void
-): [() => void, () => void] => {
-  const handler = (e: KeyboardEvent) => {
-    if (e.key === 'Escape') {
-      e.stopPropagation();
-      cb();
-    }
-  };
-
-  return [
-    () => element.addEventListener('keydown', handler),
-    () => element.removeEventListener('keydown', handler),
-  ];
-};
-
-export const onEscHandler = (element: React.RefObject<HTMLElement>, isEnabled: boolean, cb: () => void) => () => {
-  const el = element && element.current;
-  if (!el) { return; }
-
-  const [addEvListener, removeEvListener] = onEsc(el, cb);
-  if (isEnabled) {
-    addEvListener();
-  }
-
-  return removeEvListener;
-};
-
-export const useOnEsc = (element: React.RefObject<HTMLElement>, isEnabled: boolean, cb: () => void) => {
-  React.useEffect(onEscHandler(element, isEnabled, cb), [element, isEnabled]);
-};
